Square() is now replacing BoardIterator, so we can now get everything we need from the Engine.Board.
Both the UI Board and the Engine board are now coupled so we have a way to keep them synchronized.

+++ source/UI/ChessGrid.cs
@@ -22,8 +22,8 @@ namespace SKChess
 
         private void ChessGrid_Load(object sender, EventArgs e)
         {
-            int rows = 8;
-            int columns = 8;
+            ushort rows = 8;
+            ushort columns = 8;
             int squareSize = 72;
 
             //ConfigurationManager.OpenExeConfiguration(System.Environment.CurrentDirectory
